moved handle-component-edit-mode-container to sync floder
