+++ packages/packages/core/editor-components/src/init.ts
@@ -34,7 +34,7 @@ import { InstanceEditingPanel } from './components/instance-editing-panel/instan
 import { OverridablePropControl } from './components/overridable-props/overridable-prop-control';
 import { OverridablePropIndicator } from './components/overridable-props/overridable-prop-indicator';
 import { COMPONENT_WIDGET_TYPE, createComponentType } from './create-component-type';
-import { initHandleComponentEditModeContainer } from './hooks/handle-component-edit-mode-container';
+import { initHandleComponentEditModeContainer } from './sync/handle-component-edit-mode-container';
 import { initMcp } from './mcp';
 import { PopulateStore } from './populate-store';
 import { initCircularNestingPrevention } from './prevent-circular-nesting';
