+++ packages/packages/core/editor-components/src/sync/handle-component-edit-mode-container.ts
